requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.2.0
requests>=2.31.0
google-auth>=2.28.0

pytest>=8.0.0
httpx>=0.27.0
